home screen: repaired inconsistent animations
Created new branch to investigate leveraging a coordinator layout. By
allowing it to coordinate the major animations, motionLayout could focus
on the header animations. This reduced the need for the motion scene to
manage visibility of so many things (empty views v full views). The end
result is the animations behave more consistently and ghosting of hidden
views is no longer occuring. There is still a minor issue with the
account balance flickering while it animates but I suspect this is a bug
in Android. I will create a minimum example to prove it and file a bug
report.

Squashed commit of the following:

commit 62eb2138f2a0f908b9b4768eefeca6cf3f7ef267
commit 6bc77946a34a2a7e83fff061b7263d0f264589ee
commit 383fe6995778cc1545682cd0c164700a13a35775
commit 33a47eff4ec9ba64b1904d8ae9d9e0032f73b9bb
commit 5c7f85415d1dba5221a14e5a53ca90327c549afe
commit 4f77d24bb7ddf5f10cda827ac1603e71b10b02a6
commit c022dd18a2b3ba9bfd6bae853d1079b72093ca9b
commit bc3ec349d260b4c6b1256c26dcdf5e7cf9500501